Move `ResourceMemory` away from the `suballocator` module
