+++ vulkano/src/buffer/subbuffer.rs
@@ -301,7 +301,7 @@ where
     /// 64. [`SubbufferAllocator`] does this automatically.
     ///
     /// [host-coherent]: crate::memory::MemoryPropertyFlags::HOST_COHERENT
-    /// [`invalidate_range`]: crate::memory::allocator::ResourceMemory::invalidate_range
+    /// [`invalidate_range`]: crate::memory::ResourceMemory::invalidate_range
     /// [`non_coherent_atom_size`]: crate::device::Properties::non_coherent_atom_size
     /// [`write`]: Self::write
     /// [`SubbufferAllocator`]: super::allocator::SubbufferAllocator
@@ -387,7 +387,7 @@ where
     /// does this automatically.
     ///
     /// [host-coherent]: crate::memory::MemoryPropertyFlags::HOST_COHERENT
-    /// [`flush_range`]: crate::memory::allocator::ResourceMemory::flush_range
+    /// [`flush_range`]: crate::memory::ResourceMemory::flush_range
     /// [`non_coherent_atom_size`]: crate::device::Properties::non_coherent_atom_size
     /// [`read`]: Self::read
     /// [`SubbufferAllocator`]: super::allocator::SubbufferAllocator
@@ -1140,9 +1140,9 @@ mod tests {
         memory::{
             allocator::{
                 AllocationCreateInfo, AllocationType, DeviceLayout, MemoryAllocator,
-                ResourceMemory, StandardMemoryAllocator,
+                StandardMemoryAllocator,
             },
-            MemoryRequirements,
+            MemoryRequirements, ResourceMemory,
         },
     };
